+++ examples/mcp-server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # mcp-server-example
 
+## 0.0.1-alpha.3
+
+### Patch Changes
+
+- Updated dependencies [[`fa25629`](https://github.com/mattppal/mcut/commit/fa256296c4f284c58f289fc448564fc9c2fe2b88)]:
+  - @mcut/mcp-server@0.1.0-alpha.4
+
+## 0.0.1-alpha.2
+
+### Patch Changes
+
+- Updated dependencies [[`9b17ac7`](https://github.com/mattppal/mcut/commit/9b17ac71a1314a9197d350ae110783537b13243a)]:
+  - @mcut/timeline@0.1.0-alpha.3
+  - @mcut/mcp-server@0.1.0-alpha.3
+
 ## 0.0.1-alpha.1
 
 ### Patch Changes